// superset-frontend/plugins/superset-plugin-chart-leaflet-map/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import {
  CategoricalColorScale,
  DataRecordFilters,
  FilterState,
  HandlerFunction,
  QueryFormData,
  SetDataMaskHook,
  supersetTheme,
  TimeseriesDataRecord,
} from '@superset-ui/core';

export interface SupersetPluginChartLeafletMapStylesProps {
  height: number;
  width: number;
  headerFontSize: keyof typeof supersetTheme.typography.sizes;
  boldText: boolean;
}

interface SupersetPluginChartLeafletMapCustomizeProps {
  headerText: string;
}

export type SupersetPluginChartLeafletMapQueryFormData = QueryFormData &
  SupersetPluginChartLeafletMapStylesProps &
  SupersetPluginChartLeafletMapCustomizeProps;

export type SupersetPluginChartLeafletMapProps =
  SupersetPluginChartLeafletMapStylesProps &
    SupersetPluginChartLeafletMapCustomizeProps & {
      data: TimeseriesDataRecord[];
      emitCrossFilters: boolean | undefined;
      setDataMask: SetDataMaskHook;
      onContextMenu: HandlerFunction;
      filterState: FilterState;
      filters?: DataRecordFilters;
      colorScheme: any;
      colorFn: any;
      sliceId: any;
      // add typing here for the props you pass in from transformProps.ts!
    };